[Client] chore: ItemsContext 에서 StatusContext로 이름 변경 및 게임 레벨, 속도에 대한 정보를 같이 관리할 수 있도록 변경
