+++ client/client.js
@@ -2,15 +2,15 @@ import React from 'react';
 import { render } from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 
-import { ItemsProvider } from '@contexts/items';
+import { StatusProvider } from '@contexts/status';
 
 import App from '@layouts/App';
 
 render(
   <BrowserRouter>
-    <ItemsProvider>
+    <StatusProvider>
       <App />
-    </ItemsProvider>
+    </StatusProvider>
   </BrowserRouter>,
   document.getElementById('app'),
 );
